requirements.txt
llama-api-client>=0.1.0
numpy>=1.24.0
PyPDF2>=3.0.0
python-dotenv>=0.19.0
qdrant-client>=1.7.0
scikit_learn==1.6.1
sentence-transformers>=2.2.2
